refactor: split _builder.py into _execution_providers + _optimizations
_builder.py (914 lines) served three distinct concerns. Split into:

- _execution_providers.py (~232 lines): EpCapabilities dataclass,
  EpRegistry class with dict-compatible API, ep_registry singleton,
  register_ep() / get_ep() helpers, all 6 built-in EP definitions in
  _register_builtins(). Standalone — depends only on onnx_ir.

- _optimizations.py (~461 lines): SymbolicShapeInferencePass,
  CleanupMetadataPass, _DEFAULT_PASSES, trace infra (_count_ops,
  _TraceEntry, etc.), _get_optimization_passes(), optimize_model()
  (public alias: _optimize). Imports from _execution_providers + _flags.

- _builder.py (~416 lines): build() + build_from_module() + resolve_dtype()
  + dtype helpers only. Re-exports EpCapabilities, _EP_REGISTRY,
  SymbolicShapeInferencePass, CleanupMetadataPass, _count_ops,
  _count_all_ops for zero-breakage backward compat.

_EP_REGISTRY is now a shim pointing to ep_registry so all existing
callers (.get(), frozenset(), sorted()) continue to work unchanged.

Consumer updates:
- _ep_validation.py: imports ep_registry from _execution_providers;
  KNOWN_EPS derived from ep_registry.names(); adds register_unsupported_combo()
- integrations/ort_genai/ep_config.py: imports ep_registry from _execution_providers
- tests/ep_optimization_test.py: trace caplog targets mobius._optimizations

All 2326 tests pass (2274 build_graph + src, 52 EP tests).
